refactor: Split python-bindings/lib.rs into modular structure
- Extract type conversions (sqlvalue_to_py, py_to_sqlvalue, param binding) into conversions.rs
- Extract Database connection class into connection.rs
- Extract Cursor class and query execution logic into cursor.rs
- Keep profiling utilities in existing profiling.rs module
- Update lib.rs to declare modules, export public types, and maintain module-level documentation

The refactoring improves maintainability by separating concerns:
- conversions.rs handles Python ↔ Rust type conversions
- connection.rs manages database connections
- cursor.rs handles statement execution, caching, and result fetching
- profiling.rs provides performance profiling utilities

All existing tests pass and public API remains unchanged.
